requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
apprise==1.9.3
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==3.0.0
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coolname==2.2.0
croniter==5.0.1
cryptography==45.0.4
curl_cffi==0.11.4
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.1
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.4
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.5.0
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.3.0
oauthlib==3.3.0
opentelemetry-api==1.34.1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandas-gbq==0.29.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peewee==3.18.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
pinecone==7.2.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
plotly==6.1.2
prefect==3.4.7
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.0.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
streamlit-image-zoom==0.0.4
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.13.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.5.0
uv==0.7.13
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.14
yarl==1.20.1
yfinance==0.2.63
zipp==3.23.0
zstandard==0.23.0
